+++ community/src/main/java/com/lawencon/community/pojo/paymentTransaction/PojoUpdatePaymentTransactionReq.java
@@ -6,7 +6,7 @@ public class PojoUpdatePaymentTransactionReq {
 	@NotNull(message = "Id cannot be empty")
 	private String id;
 	@NotNull(message = "Version cannot be empty")
-	private Long version;
+	private Integer version;
 	@NotNull(message = "is active must be true or false")
 	private Boolean isActive;
